+++ apps/storybook-icons/utils/react-icons.ts
@@ -100,8 +100,10 @@ import { LaunchSmallIcon } from '@utilitywarehouse/hearth-react-icons/lib/Launch
 import { LinkMediumIcon } from '@utilitywarehouse/hearth-react-icons/lib/LinkMediumIcon';
 import { LinkSmallIcon } from '@utilitywarehouse/hearth-react-icons/lib/LinkSmallIcon';
 import { LinkedInMediumIcon } from '@utilitywarehouse/hearth-react-icons/lib/LinkedInMediumIcon';
+import { ListSmallIcon } from '@utilitywarehouse/hearth-react-icons/lib/ListSmallIcon';
 import { LocateSmallIcon } from '@utilitywarehouse/hearth-react-icons/lib/LocateSmallIcon';
 import { LocationMediumIcon } from '@utilitywarehouse/hearth-react-icons/lib/LocationMediumIcon';
+import { LocationSmallIcon } from '@utilitywarehouse/hearth-react-icons/lib/LocationSmallIcon';
 import { LockMediumIcon } from '@utilitywarehouse/hearth-react-icons/lib/LockMediumIcon';
 import { LockSmallIcon } from '@utilitywarehouse/hearth-react-icons/lib/LockSmallIcon';
 import { LoginMediumIcon } from '@utilitywarehouse/hearth-react-icons/lib/LoginMediumIcon';
@@ -285,8 +287,10 @@ export const reactIcons = [
   LinkMediumIcon,
   LinkSmallIcon,
   LinkedInMediumIcon,
+  ListSmallIcon,
   LocateSmallIcon,
   LocationMediumIcon,
+  LocationSmallIcon,
   LockMediumIcon,
   LockSmallIcon,
   LoginMediumIcon,