+++ packages/react/src/components/ProgressStepper/ProgressStep.context.ts
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+import { ProgressStepStatus } from './ProgressStep.props';
+
+export type ProgressStepContextOptions = {
+  status: ProgressStepStatus;
+};
+
+export const ProgressStepContext = createContext<ProgressStepContextOptions | null>(null);